nopCommerce Dev to Freemium Master

+++ src/Presentation/Nop.Web/Controllers/BoardsController.cs
@@ -4,9 +4,10 @@ using System.Linq;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Nop.Core;
-using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Forums;
+using Nop.Core.Domain.Security;
 using Nop.Core.Rss;
+using Nop.Services.Customers;
 using Nop.Services.Forums;
 using Nop.Services.Localization;
 using Nop.Web.Factories;
@@ -18,11 +19,14 @@ using Nop.Web.Models.Boards;
 namespace Nop.Web.Controllers
 {
     [HttpsRequirement(SslRequirement.No)]
+    [AutoValidateAntiforgeryToken]
     public partial class BoardsController : BasePublicController
     {
         #region Fields
 
+        private readonly CaptchaSettings _captchaSettings;
         private readonly ForumSettings _forumSettings;
+        private readonly ICustomerService _customerService;
         private readonly IForumModelFactory _forumModelFactory;
         private readonly IForumService _forumService;
         private readonly ILocalizationService _localizationService;
@@ -34,7 +38,9 @@ namespace Nop.Web.Controllers
 
         #region Ctor
 
-        public BoardsController(ForumSettings forumSettings,
+        public BoardsController(CaptchaSettings captchaSettings,
+            ForumSettings forumSettings,
+            ICustomerService customerService,
             IForumModelFactory forumModelFactory,
             IForumService forumService,
             ILocalizationService localizationService,
@@ -42,7 +48,9 @@ namespace Nop.Web.Controllers
             IWebHelper webHelper,
             IWorkContext workContext)
         {
+            _captchaSettings = captchaSettings;
             _forumSettings = forumSettings;
+            _customerService = customerService;
             _forumModelFactory = forumModelFactory;
             _forumService = forumService;
             _localizationService = localizationService;
@@ -187,6 +195,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult ForumWatch(int id)
         {
             var watchTopic = _localizationService.GetResource("Forum.WatchForum");
@@ -248,6 +257,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult TopicWatch(int id)
         {
             var watchTopic = _localizationService.GetResource("Forum.WatchTopic");
@@ -300,8 +310,7 @@ namespace Nop.Web.Controllers
             return View(model);
         }
 
-        [HttpPost]
-        [PublicAntiForgery]
+        [HttpPost]        
         public virtual IActionResult TopicMove(TopicMoveModel model)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -322,7 +331,6 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         public virtual IActionResult TopicDelete(int id)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -372,8 +380,8 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
-        public virtual IActionResult TopicCreate(EditForumTopicModel model)
+        [ValidateCaptcha]
+        public virtual IActionResult TopicCreate(EditForumTopicModel model, bool captchaValid)
         {
             if (!_forumSettings.ForumsEnabled)
                 return RedirectToRoute("Homepage");
@@ -381,6 +389,12 @@ namespace Nop.Web.Controllers
             var forum = _forumService.GetForumById(model.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
+
+            //validate CAPTCHA
+            if (_captchaSettings.Enabled && _captchaSettings.ShowOnForum && !captchaValid)
+            {
+                ModelState.AddModelError("", _localizationService.GetResource("Common.WrongCaptchaMessage"));
+            }
 
             if (ModelState.IsValid)
             {
@@ -490,8 +504,8 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
-        public virtual IActionResult TopicEdit(EditForumTopicModel model)
+        [ValidateCaptcha]
+        public virtual IActionResult TopicEdit(EditForumTopicModel model, bool captchaValid)
         {
             if (!_forumSettings.ForumsEnabled)
                 return RedirectToRoute("Homepage");
@@ -500,9 +514,16 @@ namespace Nop.Web.Controllers
 
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
-            var forum = forumTopic.Forum;
+
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
+
+            //validate CAPTCHA
+            if (_captchaSettings.Enabled && _captchaSettings.ShowOnForum && !captchaValid)
+            {
+                ModelState.AddModelError("", _localizationService.GetResource("Common.WrongCaptchaMessage"));
+            }
 
             if (ModelState.IsValid)
             {
@@ -599,11 +620,11 @@ namespace Nop.Web.Controllers
 
             //redisplay form
             _forumModelFactory.PrepareTopicEditModel(forumTopic, model, true);
+
             return View(model);
         }
 
         [HttpPost]
-        [PublicAntiForgery]
         public virtual IActionResult PostDelete(int id)
         {
             if (!_forumSettings.ForumsEnabled)
@@ -613,26 +634,28 @@ namespace Nop.Web.Controllers
                 });
 
             var forumPost = _forumService.GetPostById(id);
-            if (forumPost != null)
+
+            //TODO: ???? get topic one more time because it can be deleted (first or only post deleted)
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
+
+            if (forumPost != null && forumTopic != null)
             {
                 if (!_forumService.IsCustomerAllowedToDeletePost(_workContext.CurrentCustomer, forumPost))
                     return Challenge();
 
-                var forumTopic = forumPost.ForumTopic;
-                var forumId = forumTopic.Forum.Id;
-                var forumSlug = _forumService.GetForumSeName(forumTopic.Forum);
+                var forum = _forumService.GetForumById(forumTopic.ForumId);
+                var forumSlug = _forumService.GetForumSeName(forum);
 
-                _forumService.DeletePost(forumPost);
-
-                //get topic one more time because it can be deleted (first or only post deleted)
-                forumTopic = _forumService.GetTopicById(forumPost.TopicId);
                 if (forumTopic == null)
                 {
                     return Json(new
                     {
-                        redirect = Url.RouteUrl("ForumSlug", new { id = forumId, slug = forumSlug }),
+                        redirect = Url.RouteUrl("ForumSlug", new { id = forum.Id, slug = forumSlug }),
                     });
-                }
+                }                
+
+                _forumService.DeletePost(forumPost);
+                                
                 return Json(new
                 {
                     redirect = Url.RouteUrl("TopicSlug", new { id = forumTopic.Id, slug = _forumService.GetTopicSeName(forumTopic) }),
@@ -662,8 +685,8 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
-        public virtual IActionResult PostCreate(EditForumPostModel model)
+        [ValidateCaptcha]
+        public virtual IActionResult PostCreate(EditForumPostModel model, bool captchaValid)
         {
             if (!_forumSettings.ForumsEnabled)
                 return RedirectToRoute("Homepage");
@@ -671,6 +694,12 @@ namespace Nop.Web.Controllers
             var forumTopic = _forumService.GetTopicById(model.ForumTopicId);
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
+
+            //validate CAPTCHA
+            if (_captchaSettings.Enabled && _captchaSettings.ShowOnForum && !captchaValid)
+            {
+                ModelState.AddModelError("", _localizationService.GetResource("Common.WrongCaptchaMessage"));
+            }
 
             if (ModelState.IsValid)
             {
@@ -730,13 +759,13 @@ namespace Nop.Web.Controllers
 
                     var pageSize = _forumSettings.PostsPageSize > 0 ? _forumSettings.PostsPageSize : 10;
 
-                    var pageIndex = (_forumService.CalculateTopicPageIndex(forumPost.TopicId, pageSize, forumPost.Id) + 1);
+                    var pageIndex = _forumService.CalculateTopicPageIndex(forumPost.TopicId, pageSize, forumPost.Id) + 1;
                     var url = string.Empty;
                     if (pageIndex > 1)
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     else
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
-                    return Redirect($"{url}#{forumPost.Id}");
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
+                    return LocalRedirect($"{url}#{forumPost.Id}");
                 }
                 catch (Exception ex)
                 {
@@ -745,7 +774,8 @@ namespace Nop.Web.Controllers
             }
 
             //redisplay form
-            _forumModelFactory.PreparePostCreateModel(forumTopic, 0, true);
+            model = _forumModelFactory.PreparePostCreateModel(forumTopic, 0, true);
+
             return View(model);
         }
 
@@ -766,8 +796,8 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost]
-        [PublicAntiForgery]
-        public virtual IActionResult PostEdit(EditForumPostModel model)
+        [ValidateCaptcha]
+        public virtual IActionResult PostEdit(EditForumPostModel model, bool captchaValid)
         {
             if (!_forumSettings.ForumsEnabled)
                 return RedirectToRoute("Homepage");
@@ -779,13 +809,19 @@ namespace Nop.Web.Controllers
             if (!_forumService.IsCustomerAllowedToEditPost(_workContext.CurrentCustomer, forumPost))
                 return Challenge();
 
-            var forumTopic = forumPost.ForumTopic;
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
 
-            var forum = forumTopic.Forum;
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
+
+            //validate CAPTCHA
+            if (_captchaSettings.Enabled && _captchaSettings.ShowOnForum && !captchaValid)
+            {
+                ModelState.AddModelError("", _localizationService.GetResource("Common.WrongCaptchaMessage"));
+            }
 
             if (ModelState.IsValid)
             {
@@ -837,13 +873,13 @@ namespace Nop.Web.Controllers
                     var url = string.Empty;
                     if (pageIndex > 1)
                     {
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     }
                     else
                     {
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
                     }
-                    return Redirect($"{url}#{forumPost.Id}");
+                    return LocalRedirect($"{url}#{forumPost.Id}");
                 }
                 catch (Exception ex)
                 {
@@ -852,17 +888,18 @@ namespace Nop.Web.Controllers
             }
 
             //redisplay form
-            _forumModelFactory.PreparePostEditModel(forumPost, true);
+            model = _forumModelFactory.PreparePostEditModel(forumPost, true);
+
             return View(model);
         }
 
         public virtual IActionResult Search(string searchterms, bool? adv, string forumId,
-            string within, string limitDays, int page = 1)
+            string within, string limitDays, int pageNumber = 1)
         {
             if (!_forumSettings.ForumsEnabled)
                 return RedirectToRoute("Homepage");
 
-            var model = _forumModelFactory.PrepareSearchModel(searchterms, adv, forumId, within, limitDays, page);
+            var model = _forumModelFactory.PrepareSearchModel(searchterms, adv, forumId, within, limitDays, pageNumber);
             return View(model);
         }
 
@@ -876,6 +913,7 @@ namespace Nop.Web.Controllers
         }
 
         [HttpPost, ActionName("CustomerForumSubscriptions")]
+        [IgnoreAntiforgeryToken]
         public virtual IActionResult CustomerForumSubscriptionsPOST(IFormCollection formCollection)
         {
             foreach (var key in formCollection.Keys)
@@ -885,7 +923,7 @@ namespace Nop.Web.Controllers
                 if (value.Equals("on") && key.StartsWith("fs", StringComparison.InvariantCultureIgnoreCase))
                 {
                     var id = key.Replace("fs", "").Trim();
-                    if (int.TryParse(id, out int forumSubscriptionId))
+                    if (int.TryParse(id, out var forumSubscriptionId))
                     {
                         var forumSubscription = _forumService.GetSubscriptionById(forumSubscriptionId);
                         if (forumSubscription != null && forumSubscription.CustomerId == _workContext.CurrentCustomer.Id)
@@ -909,7 +947,7 @@ namespace Nop.Web.Controllers
             if (forumPost == null)
                 return new NullJsonResult();
 
-            if (!_workContext.CurrentCustomer.IsRegistered())
+            if (!_customerService.IsRegistered(_workContext.CurrentCustomer))
                 return Json(new
                 {
                     Error = _localizationService.GetResource("Forum.Votes.Login"),
